+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -455,6 +455,7 @@ sub build_isf_public_interface_contract {
             't/1286-isf-aggregate-condition-values.t',
             't/1287-isf-aggregate-drive-values.t',
             't/1288-isf-aggregate-drive-expression-values.t',
+            't/1289-isf-aggregate-drive-call-values.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',